requirements.txt
torch==2.2.0
torchvision==0.17.0
timm==0.9.12
transformers==4.36.0
pandas==2.1.4
numpy==1.26.3
Pillow==10.2.0
albumentations==2.0.8
opencv-python==4.9.0.80
scikit-learn==1.3.2
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.2
torchmetrics==1.3.1
jupyter==1.0.0
jupyterlab>=3.0.0
ipykernel>=6.0.0